styled search bar and added it to nav, styled footer links to display their images instead of titles. Show route is giving me an error about Meta Data? and search is giving an error about filter? help pls my brain hurts
