+++ src/App.css
@@ -32,6 +32,18 @@ font-size: 25px;
 .logo:hover{
  text-decoration: none;
 }
+
+/* search bar styling  */
+.submitBtn{
+  max-width: 30px;
+  max-height: 25px;
+  position: relative;top: -20px;
+}
+input[type = search]{
+  position: relative; top:-20px;
+
+}
+
 /* landing styling  */
 .landing{
   color: whitesmoke;
@@ -84,7 +96,7 @@ input[type=text]{
   margin: 5px 20px 10px 20px;
   border: none;
 }
-input[type=submit]{
+input[type=submit]:not(.submitBtn){
   font-family: 'PT Serif', serif;
   font-size: 18px;
   padding: 7px;
@@ -172,11 +184,25 @@ label{
   font-size: 18px;
   font-weight: bold;
 }
+
 /* footer styling */
 .footer{
   margin-top: 60px;
   text-align:center;
   color: whitesmoke;
+}
+.footerTag{
+  margin-top: 0px;
+}
+.socials{
+  display: flex;
+  flex-wrap: wrap;
+ justify-content: center;
+}
+.twitter, .youtube, .contacts{
+  max-width: 50px;
+  border:solid white;
+  border-radius: 50%;
 }
 .account {
   color: white
