+++ src/components/Header.js
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+import Searchbar from "./Search";
 
 const Header = (props) => {
   return (
@@ -8,24 +9,19 @@ const Header = (props) => {
      
       <nav className="navBar">
 
-        <Link to="/">
-          <div>Home</div>
-        </Link>
         <Link to="/about">
-            <div>About</div>
-        </Link>
-        <Link to="/search">
-          <div>Search</div>
-        </Link>
-        <Link to="/" className='logo'>
-          <div className='logo'>JxFUNDS</div>
+          <div>About</div>
         </Link>
         <Link to="/jxfunds">
-          <div>Funds</div>
+            <div>Funds</div>
         </Link>
         <Link to="/jxfunds/create">
           <div>Create</div>
         </Link>
+        <Link to="/" className='logo'>
+          <div className='logo'>JxFUNDS</div>
+        </Link>
+        <Searchbar />
         <Link to="/account">
           <div>Account</div>
         </Link>
